Fixed warnings.

+++ MainGame/TerrainObject.h
@@ -4,12 +4,12 @@
 class TerrainObject : public StaticObject
 {
 protected:
-	Vector2f focus1, focus2;
+	Vector2i focus1, focus2;
 	int ellipseSize;
 public:
 	TerrainObject(std::string objectName, Vector2f centerPosition);
-	Vector2f getFocus1() const { return focus1; }
-	Vector2f getFocus2() const { return focus2; }
+	Vector2i getFocus1() const { return focus1; }
+	Vector2i getFocus2() const { return focus2; }
 	int getEllipseSize() const { return ellipseSize; }
 };
